+++ packages/ispect_jira/example/lib/main.dart
@@ -73,7 +73,7 @@ class _AppState extends State<App> {
           options: ISpectOptions(
             locale: locale,
             panelButtons: [
-              (
+              ISpectPanelButtonItem(
                 icon: Icons.copy_rounded,
                 label: 'Token',
                 onTap: (context) {
@@ -82,7 +82,7 @@ class _AppState extends State<App> {
               ),
             ],
             panelItems: [
-              (
+              ISpectPanelItem(
                 icon: Icons.home,
                 enableBadge: false,
                 onTap: (context) {
@@ -91,7 +91,7 @@ class _AppState extends State<App> {
               ),
             ],
             actionItems: [
-              ISpectifyActionItem(
+              ISpectActionItem(
                 title: 'ISpect',
                 icon: Icons.bug_report_outlined,
                 onTap: (context) {